+++ Lab5/RectangleLawn.cpp
@@ -45,20 +45,20 @@ namespace lab5
 
 	unsigned int RectangleLawn::GetFencePrice(eFenceType fenceType) const
 	{
-		unsigned int fencePrice = 0;
+		float fencePrice = 0;
 
 		switch (fenceType)
 		{
 		case 0:
-			fencePrice = GetMinimumFencesCount() * 1.5;
+			fencePrice = static_cast<float>(GetMinimumFencesCount()) * 1.5f;
 			break;
 
 		case 1:
-			fencePrice = GetMinimumFencesCount() * 1.75;
+			fencePrice = static_cast<float>(GetMinimumFencesCount()) * 1.75f;
 			break;
 
 		}
 
-		return fencePrice;
+		return static_cast<unsigned int>(fencePrice);
 	}
 }
